+++ src/Model/ImportedShlinkOrphanVisit.php
@@ -6,15 +6,15 @@ namespace Shlinkio\Shlink\Importer\Model;
 
 use DateTimeInterface;
 
-final class ImportedShlinkOrphanVisit
+final readonly class ImportedShlinkOrphanVisit
 {
     public function __construct(
-        public readonly string $referer,
-        public readonly string $userAgent,
-        public readonly DateTimeInterface $date,
-        public readonly string $visitedUrl,
-        public readonly string $type,
-        public readonly ImportedShlinkVisitLocation|null $location,
+        public string $referer,
+        public string $userAgent,
+        public DateTimeInterface $date,
+        public string $visitedUrl,
+        public string $type,
+        public ImportedShlinkVisitLocation|null $location,
     ) {
     }
 }